requirements.txt
aiohttp
python-telegram-bot==20.3
fastapi
uvicorn
